realtime-whatsmart-shop/requirements.txt
Flask==2.3.3
twilio==8.2.2
redis==4.5.5
opencv-python-headless==4.8.0.76
numpy==1.26.4
requests==2.31.0
openai-whisper==20231117
stripe==5.5.0
mediapipe==0.10.14
